// resources/js/admin/game-maps/definitions/game-map-form-errors-definition.ts
export default interface GameMapFormErrorsDefinition {
  name?: string;
  description?: string;
  kingdom_color?: string;
  map?: string;
  replacement_image_acknowledged?: string;
  xp_bonus?: string;
  skill_training_bonus?: string;
  drop_chance_bonus?: string;
  enemy_stat_bonus?: string;
  character_attack_reduction?: string;
  required_location_id?: string;
}